# RFC: Try implement callsite cache for runtime dynamic generic function call