--- a/src/jltypes.c
+++ b/src/jltypes.c
@@ -3430,9 +3430,10 @@ void jl_init_types(void)
     jl_lambda_info_type =
         jl_new_datatype(jl_symbol("LambdaStaticData"),
                         jl_any_type, jl_emptysvec,
-                        jl_svec(15, jl_symbol("ast"), jl_symbol("sparams"),
+                        jl_svec(16, jl_symbol("ast"), jl_symbol("sparams"),
                                 jl_symbol("tfunc"), jl_symbol("name"),
                                 jl_symbol("roots"),
+                                jl_symbol("dynRoots"),
                                 /* jl_symbol("specTypes"),
                                    jl_symbol("unspecialized"),
                                    jl_symbol("specializations")*/
@@ -3442,14 +3443,16 @@ void jl_init_types(void)
                                 jl_symbol("file"), jl_symbol("line"),
                                 jl_symbol("inferred"),
                                 jl_symbol("pure")),
-                        jl_svec(15, jl_any_type, jl_simplevector_type,
+                        jl_svec(16, jl_any_type, jl_simplevector_type,
                                 jl_any_type, jl_sym_type,
-                                jl_any_type, jl_any_type,
-                                jl_any_type, jl_array_any_type,
+                                jl_any_type,
+                                jl_any_type,
+                                jl_any_type, jl_any_type, jl_array_any_type,
                                 jl_module_type, jl_any_type,
                                 jl_any_type,
                                 jl_sym_type, jl_int32_type,
-                                jl_bool_type, jl_bool_type),
+                                jl_bool_type,
+                                jl_bool_type),
                         0, 1, 4);
 
     jl_box_type =
@@ -3477,7 +3480,7 @@ void jl_init_types(void)
                         0, 1, 0);
 
     jl_svecset(jl_method_type->types, 4, jl_function_type);
-    jl_svecset(jl_lambda_info_type->types, 6, jl_function_type);
+    jl_svecset(jl_lambda_info_type->types, 7, jl_function_type);
 
     jl_bottom_func = jl_new_closure(jl_f_no_function, (jl_value_t*)jl_emptysvec, NULL);