This is my attempt to implement @carnaval 's JSoC idea.

It has a few hacky point (rooting function objects, the way it gives up on types as argument).

The current implementation shows a speed up of ~20% for a no-op function call and is about 7x slower than a direct non-inlined callsite using the following [benchmark](https://github.com/yuyichao/explore/blob/master/julia/callsite_cache/bench.jl)

``` julia
@noinline f1(a::ANY, b::ANY) = nothing
@noinline f2(a::ANY, b::ANY) = a + b

k1(a::ANY, b::ANY, n) = for i in 1:n
    f1(a, b)
end

k2(a::ANY, b::ANY, n) = for i in 1:n
    f2(a, b)
end

@noinline f3(a, b) = a + b

k3(a::ANY, b::ANY, n) = for i in 1:n
    f3(a, b)
end

k4(a, b, n) = for i in 1:n
    f3(a, b)
end

k5(n) = for i in 1:n
    f2(1, 2)
end

k1(1, 2, 3)
k2(1, 2, 3)
k3(1, 2, 3)
k4(1, 2, 3)
k5(3)

@time k1(1, 2, 100_000_000)
@time k2(1, 2, 100_000_000)
@time k3(1, 2, 100_000_000)
@time k4(1, 2, 100_000_000)
@time k5(100_000_000)

# Profile.clear()
# @profile k3(1, 2, 300_000_000)
# Profile.print(C=true)
```

Note that `k4` is a direct callsite version of `k3` for comparason.

Output on master

``` julia
   1.502 seconds      (156 allocations: 10760 bytes)
   3.172 seconds      (4 allocations: 144 bytes)
   1.515 seconds      (4 allocations: 144 bytes)
 175.886 milliseconds (4 allocations: 144 bytes)
   2.496 seconds      (4 allocations: 144 bytes)
```

With this PR

``` julia
   1.173 seconds      (156 allocations: 10760 bytes)
   2.571 seconds      (4 allocations: 144 bytes)
   1.256 seconds      (4 allocations: 144 bytes)
 183.999 milliseconds (4 allocations: 144 bytes)
   1.709 seconds      (4 allocations: 144 bytes)
```

Unfortunately (for fortunately) the base is reasonably well written so the speedup in the test is not very visible (although it also doesn't show any slowdown). This is also my first time to write any code that is very cache sensitive so suggestions are definitely welcome.

@JeffBezanson