@yuyichao 2015-06-27
I think this is probably the first time I see a OOM error on 32bit Travis https://travis-ci.org/JuliaLang/julia/jobs/68377025