@yuyichao 2015-11-06
I'm rebasing it on the codegen rewrite and want to clear out the staled commit.... The original commit is backed up on a different (local) branch. I didn't know pushing the clean up out cause github to think the PR is merged .............................................. anyway, will either reopen or open a different one when I'm ready....