@yuyichao 2015-11-06
I did realize that (although I'm still interested in what kind of performance I can get by implementing some cache check inline).

Do you have a plan for #13412 regarding codegen. e.g. `cfunction` (especially on closure) clean up of different types etc?