@JeffBezanson 2015-11-06
@vtjnash and I have been discussing closure support in `cfunction` using http://llvm.org/docs/LangRef.html#trampoline-intrinsics. Non-closures are now "ghost" objects with 0 size, so they can be omitted from argument lists, allowing the existing uses of `cfunction` to keep working unmodified.