@yuyichao 2015-11-08
I rebase (more like reimplement) this PR on the current master. The cache match is inlined in codegen now (the `memcmp` takes most of the time in the previous version). This gives a significant speed up compare to the original version and for the simple case this is now ~ 3 times faster than master.

This could in principle make #265 worse since there's more caching now. Luckily, this is a relatively easier one to solve since the cache invalidation is cheap (e.g. if we add an age to generic function).

There's probably still some improvement to be implemented and I also haven't checked for regression yet...