@yuyichao 2015-11-08
Forgot to add benchmark numbers... Using the microbenchmark above.

Current master
```
  1.422174 seconds (154 allocations: 10.479 KB)
  2.949819 seconds (4 allocations: 160 bytes)
  1.432361 seconds (4 allocations: 160 bytes)
  0.162515 seconds (4 allocations: 160 bytes)
  2.395927 seconds (4 allocations: 160 bytes)
```

This PR.
```
  0.454639 seconds (154 allocations: 10.479 KB)
  1.031668 seconds (4 allocations: 160 bytes)
  0.690012 seconds (4 allocations: 160 bytes)
  0.176748 seconds (4 allocations: 160 bytes)
  0.949349 seconds (4 allocations: 160 bytes)
```