@JeffBezanson 2015-11-08
Wow, impressive!

What is the effect on code size? This seems like a lot of extra space to use for every generic call site. Is there evidence that `JL_GF_CALLSITE_CACHE_SIZE` greater than 1 is helpful?

It might be possible to do this with less code by just passing an extra pointer to `jl_apply_generic`, pointing to the call site's cache.