@yuyichao 2015-11-08
> What is the effect on code size?

After stripping, `sys.so` increases from `26M` to `27M`.

> Is there evidence that `JL_GF_CALLSITE_CACHE_SIZE` greater than `1` is helpful?

I imagine this will depend on the actual workload. I'm hoping that this can speed up code that is truely using dynamic dispatch as well with a cache size larger than one e.g. [this recent julia-user thread](https://groups.google.com/forum/?fromgroups=#!topic/julia-users/pTywc8Lcd28). I think I'll need some representitive workload to benchmark this....

> It might be possible to do this with less code by just passing an extra pointer to `jl_apply_generic`, pointing to the call site's cache.

This is basically the first version. The issue though is that the generic version is doing a lot of unnecessary works. (Yes, it was certainly much less code, especially since I don't need to hand write a loop in LLVM .....). The cache matching logic in this version use the type inference results to only compare the argument types (as well as the function being called).

When I was asking about codegen changes for #13412, I was also interested in whether there'll be any change to the jlcall protocol. For example, I'm wondering if we can (optionally?) pass in the types of the arguments separately to avoid boxing of isbits type arguments.