@JeffBezanson 2015-11-08
In #13412, so far the jlcall convention is the same. However it is only used internally for methods; the ABI for doing a normal julia function call is to just call `jl_apply_generic`. Passing types separately in jlcall seems like it will be a lot more complicated and expensive in general.

I think the less-code version of this is worth pursuing. Are you sure it's that much slower?