@yuyichao 2015-11-08
> I think the less-code version of this is worth pursuing. Are you sure it's that much slower?

Implementing that on top of the current version shouldn't be hard and I'll give it a try. I'm also wondering if a non-inlined but specialized version would be better.