@KristofferC 2018-10-18
Closed by https://github.com/JuliaLang/julia/pull/16839.